A good step.

+++ src/pixel_math.rs
@@ -26,8 +26,7 @@ lazy_static! {
         .collect::<Vec<f32>>();
 }
 
-#[allow(dead_code)]
-pub fn screen_to_board(x: f32, y: f32) -> Option<(u16, u16)> {
+pub fn screen_to_board(x: f32, y: f32) -> Option<(usize, usize)> {
     if x < MARGIN.0 || x > SCREEN_SIZE.0 - MARGIN.0 {
         None
     } else if y < MARGIN.1 || y > SCREEN_SIZE.1 - MARGIN.1 {
@@ -35,6 +34,6 @@ pub fn screen_to_board(x: f32, y: f32) -> Option<(u16, u16)> {
     } else {
         let i = (x - MARGIN.0) / POSITION_SIZE.0;
         let j = (y - MARGIN.1) / POSITION_SIZE.1;
-        Some((i as u16, j as u16))
+        Some((i as usize, j as usize))
     }
 }
